requirements.txt
opencv-python>=4.8.0,<5.0.0
numpy>=1.24.0,<2.0.0
matplotlib>=3.7.0
Pillow>=10.0.0
pyzbar>=0.1.9
imutils>=0.5.4
scikit-image>=0.21.0
qreader>=3.16.0
qrdet>=2.5.0
tensorflow>=2.13.0
keras>=2.13.0
torch>=2.8.0
torchvision>=0.23.0
ultralytics>=8.3.0
pandas>=2.3.0
requests>=2.32.0
tqdm>=4.67.0
shapely>=2.1.0

